Add some subject set picker tests
